+++ BankApp/Entities/Currency.cs
@@ -4,7 +4,7 @@ namespace BankApp.Entities;
 
 public class Currency
 {
-    [Key] public short Id { get; set; }
+    [Key] public int Id { get; set; }
     public string Code { get; set; } = default!;
     public decimal Bid { get; set; }
     public decimal Ask { get; set; }
