+++ BankApp/Services/CurrencyService/ICurrencyService.cs
@@ -5,6 +5,6 @@ namespace BankApp.Services.CurrencyService;
 public interface ICurrencyService
 {
     Task<IList<Currency>> GetAllCurrenciesAsync(CancellationToken cancellationToken = default);
-    Task<Currency> GetCurrencyByIdAsync(short currencyId, CancellationToken cancellationToken = default);
-    Task<bool> CurrencyExistsByIdAsync(short currencyId, CancellationToken cancellationToken = default);
+    Task<Currency> GetCurrencyByIdAsync(int currencyId, CancellationToken cancellationToken = default);
+    Task<bool> CurrencyExistsByIdAsync(int currencyId, CancellationToken cancellationToken = default);
 }
